find items through search when dropdown not fully visible
If a "custom vocab" is the last component of a "resource template", and the list includes between 9 and 25 items, some of them are not visible. This is because the dropdown menu is at the very bottom of the screen, so even if it's scrollable, the last few items are not visible (tested on both Firefox and Chromium).

Setting the threshold for enabling search at 5 (or seemingly, at a value lower than 8, but 5 is probably safer), attenuates this issue, as at least users can search (it may still be mildly confusing that they don't see it when scrolling the dropdown, but at least there's an intuitive fix). Presently, there is simply no way to see the last few items in a list of e.g. 15 items and select them. As a workaround, one may click on "add value" a few times to gain some screen space for the dropdown menu, but this is obviously not a sensible approach.